SCC: Separating SCC-hoist tests from general SCC tests
